requirements.txt
fastapi
uvicorn
ollama
python-dotenv
requests
fastapi
uvicorn[standard]
python-multipart
python-jose[cryptography]
passlib[bcrypt]